Put state and model into the screen
