+++ app/src/main/java/eu/kanade/presentation/manga/MangaNotesScreen.kt
@@ -13,7 +13,7 @@ import androidx.compose.ui.Modifier
 import eu.kanade.presentation.components.AppBar
 import eu.kanade.presentation.components.AppBarTitle
 import eu.kanade.presentation.manga.components.MangaNotesTextArea
-import eu.kanade.tachiyomi.ui.manga.notes.MangaNotesScreenState
+import eu.kanade.tachiyomi.ui.manga.notes.MangaNotesScreen
 import tachiyomi.i18n.MR
 import tachiyomi.presentation.core.components.material.Scaffold
 import tachiyomi.presentation.core.components.material.padding
@@ -21,7 +21,7 @@ import tachiyomi.presentation.core.i18n.stringResource
 
 @Composable
 fun MangaNotesScreen(
-    state: MangaNotesScreenState,
+    state: MangaNotesScreen.State,
     navigateUp: () -> Unit,
     onSave: (String) -> Unit,
     modifier: Modifier = Modifier,
